+++ packages/moxxmpp/lib/src/settings.dart
@@ -5,10 +5,8 @@ class ConnectionSettings {
     required this.jid,
     required this.password,
     required this.useDirectTLS,
-    required this.allowPlainAuth,
   });
   final JID jid;
   final String password;
   final bool useDirectTLS;
-  final bool allowPlainAuth;
 }